Transition external scanner to C
